Project restructuring
Moved controllers, models, routes, utils and views to ./src/
